[i2c,dv] Add directed test for the tx_stretch_ctrl feature
See the testpoint and vseq header for more detailed descriptions, but in short
this feature causes target-mode read transfers to halt and await software
intervention. The directed test checks for some descriptive components of the
testpoint, and also that the interrupt behaviour is in-line with the expectation
over the course of each transfer.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>