[pinmux,fpv] Fix type of mio_pad_attr in pinmux_assert_fpv.sv
The existing type wasn't quite right. The register we're talking about
has 14 bits of data. The pinmux_reg2hw_mio_pad_attr_mreg_t type has
those 14 bits of data (in the "q" fields), but also has 10 bits of
"qe" stuff that we don't want to track.

Tweak the types to make the types in the assignment match and hence
get their widths to match as well.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>