[pinmux,fpv] Fix value of input_disable in mask in pinmux_assert_fpv
This feeds into the MioAttrO_A assertion, which is trying to mirror
the path "from register value to output of block". Ignoring the
pinmux_strap_sampling instance, this is done in the design by masking
by the attr_warl_o output from a prim_pad_attr instance. This instance
drives the input_disable field in the mask to 1.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>