[i2c,dv] Split out Reference Model from Scoreboard
The current state of the I2C DV environment is not great. Very little of the
hardware is modelled, sequence items are used to represent vastly different
things in different contexts, and checking in the scoreboard is deliberately
limited to a subset of fields we believe to have modelled correctly. The vseq's
frequently intertwine stimulus and checking, and often reach into other parts of
the DV environment (e.g. agent's monitor, scoreboard) to reset state when our
modelling is insufficient. This is not great.

This change begins to setup a new, principled foundation for DV by tackling the
scoreboard. Currently, as often is done in OpenTitan DV, the scoreboard performs
two functions, both the intended checking component, but also acts as a
reference model to try and predict future DUT behaviour. This leads to a very
dense and overloaded module, and makes it difficult to modify while maintaining
a clean seperation.
This commit splits the modelling components of the existing scoreboard into a
seperate component "i2c_reference_model" (aka predictor), leaving the scoreboard
as a more bare-bones comparator. The scoreboard maintains a handle to the
reference model, and is responsible for calling "process_tl_access()" based on
observed TL-UL bus activity. Ideally the reference model would communicate with
the scoreboard only via TLM ports/fifo, but this is not yet implemented.

This change also routes stimulus seq_items, generated by the vseq's to drive the
agent (particularly for DUT-Target tests), first into the reference model and
then onto the scoreboard. In some cases the items are passed-on unchanged to the
scoreboard, but sometimes we make modifications. This architecture is maybe not
ideal, but does clean up what we have currently, and could be iterated on.

Additional/Related Changes
- Cleanup i2c_monitor tran_id to start counting from 0.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>