[top_earlgrey,padring] Implement muxing for IOB[0:3]
In packages with reduced pin count, not all pads can be bonded out.  The
direct-mapped pads have to be bonded out, but Quad SPI may not always be
required.  For use cases where IOB[0:3] are not bonded out but
SPI_HOST_D{2,3} and SPI_DEV_D{2,3} are bonded out but are not used
(because Quad SPI is not used), this commit adds muxes that allow using
those Quad SPI ports via IOB[0:3].

The mux selectors are currently tied off (to the direct connection);
they will be wired up as part of #23280.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>