[usb,sival] Use the harness even on CW310
Certain USB test do not technically require the harness on
the CW310 because there is no need to enable usb sensing. However,
not using also means that the harness is not tested in this configuration
in CI which can lead to breakage by mistake.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 354dfec31374f3a526d1edc9c6434c7413159d67)