[otbn,dv] Tidy up timing waiting for call stack corruption errors
The existing code looked at a stack_read_o output from the stack, but
that isn't quite what we want because it only goes high if the
instruction is actually committed. This might not happen, in which
case we'll request a stack pop but tell the stack not to commit the
change. But we still expect to see an error immediately. Make the
timing line up.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>