[rom] Update the length upper bound in the manifest
Fix #22922 for updating the length upper bound in the manifest.

Introduces a new constant `CHIP_ROM_EXT_RESIZABLE_SIZE_MAX` which is
the maximum size of the image that ROM is willing to boot (including normal
ROM_EXT and perso).

The `CHIP_ROM_EXT_SIZE_MAX` constant couldn't be changed for this purpose
since ROM_EXT relies on this value for calculating the BL0 offset.

Also, this commit includes a manifest unit test to ensure the manifest
extension getter works correctly in the large image case.

Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>