[sram_ctrl/dv] Avoid setting readback enable during access
This commit fixes the following error message: `Setting the value of field
"en" while containing register "sram_ctrl_regs_reg_block.readback" is
being accessed may result in loss of desired field value. A race condition
between threads concurrently accessing the register model is the likely
cause of the problem.`
The reason for this message was that the `sram_readback_en` task was spawned
multiple times, causing the error indicated in the message.

Closes #23328

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>