[rom] Move rom_verify to a separate lib for fuzzing.
This commit isolates the `rom_verify` function out of the main rom.c file to
minimize the dependencies needed to be set up for fuzzing.

Since the CFI table is shared between `rom_verify` and main rom, CFI table is
also moved to a common library, and CFI APIs are slightly changed to support
variable extern.

Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>