[rv_dm,dv] Avoid a fast JTAG clock
If the JTAG clock runs much faster than the system clock then things
can get a bit out of sync for DMI accesses. This is tracked in an
issue in the pulp-dbg module that we vendor.

Tightening the constraint slightly like this should avoid the issue
that led to the problem in the first place (where the system clock was
at 25MHz and the JTAG tck ran at about 100MHz).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>