[jtag,dv] Avoid driving to test-logic-reset in jtag_driver.sv
This code was supposed to run immediately after trst_n went high. That
is rather silly: we *know* we're going to be in that FSM state
already!

It turns out to really matter for rv_dm testing, where the JTAG
interface might not be connected immediately. The end result is that
it gets connected halfway through executing
drive_jtag_test_logic_reset, which leaves the driver and the TAP
out of sync.

Getting rid of this fixes things: the jtag driver will no longer start
driving signals until it is stimulated with an item to tell it to do
so.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>