[opentitantool] Restore baud rate after rescue
The rescue protocol has a new feature to increase the communication rate
above the power-on default, which greatly reduces the time it takes to
load a firmware image.

However, if a subsequent command want to e.g. use the serial console
after finishing flashing and rebooting, it will operate at its default
speed.

This change makes opentitantool also restore the speed setting of
HyperDebug, to match that of the OT chip under test.

Change-Id: I4352642e10ab40738599c4b03e3525c47cde32cf
Signed-off-by: Jes B. Klinke <jbk@chromium.org>