[manuf] enable local or KMS cert endorment key
This fixes the provisioning orchestrator script to enable using either:
1. a local key, or
2. a cloud KMS key
for endorsing certificates during personalization. Which key is used is
based on which parameter is passed to the orchestrator.py script. If the
`--ca_key_ckms_id` parameter is passed, the cloud KMS key will be used.
Similarly, if the `--ca_key_der_file` parameter is passed, the local key
will be used.

Signed-off-by: Tim Trippel <ttrippel@google.com>