[rv_dm,dv] Return more quickly on reset in rv_dm_tap_fsm_vseq
This is fixing two separate problems. Firstly, the rv_dm_tap_fsm_vseq
sequence takes quite a while, issuing back-to-back CSR accesses. The
code in wait_to_issue_reset fails as a result, because it can't find a
good time to inject one. To avoid that happening, we notice when
wait_to_issue_reset is waiting and wait a few cycles without doing a
CSR access to give it a chance.

The second failure mode is more confusing! Here, we are running the
vseq multiple times, and issuing resets after each one. A run might
look like this:

    Start vseq 0
    Issue reset
    Disable fork
    Start vseq 1
    Issue reset
    Disable fork
    ...

The problem is that vseq 0 might not have stopped because of the first
reset. We start vseq 1, but it stays queued because vseq 0 is still
running. We then do the "disable fork" for the second time, killing
the process that had queued up vseq 1. This causes quite a cryptic
error from uvm_sequencer_base:

  The task responsible for requesting a wait_for_grant on sequencer
  'uvm_test_top.env.m_jtag_agent.sequencer' for sequence
  'uvm_test_top.env.virtual_sequencer.rv_dm_tap_fsm_vseq' has been
  killed, to avoid a deadlock the sequence will be removed from the
  arbitration queues

That turns out to be reasonable! We killed the process for something
that was awaiting arbitration, so there's no way to start its
sequence.

Returning more quickly should fix this.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>