[rv_dm,dv] Lower reseed count for rv_dm_tap_fsm_rand_reset
This is quite a long test, taking a few minutes for an iteration. I
believe it was originally added to increase coverage numbers, but
there's probably no specific need for it right at the moment.

Knock the reseed count down to something more sensible as an interim
step.

Note that I claim I'm not just doing this to improve the nightly pass
numbers! On my laptop, I just ran 20 random seeds on the tip of my
"work" branch and had a pass rate of 80% (but took 15 minutes with 8
cores the four failures were caused by timeouts).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>