[rv_dm,dv] Increase rv_dm_jtag_dmi_csr_bit_bash timeout
We increase to to 150ms in the rv_dm base vseq, but this gets
overridden by the command line argument from the hjson file. The whole
point of that argument was to give a *higher* timeout, which isn't
quite what is happening! Bump up to 150ms as an initial step to match.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>