[rv_dm,dv] Constrain iteration counts in sba tl_access vseq
Many of these tests were timing out, for rather silly reasons. A local
example run had num_times=2 and num_trans=25 and this took about
150ms, the time limit that we impose on the uvm phase. That works out
as roughly 3ms per transaction.

Rather than bumping the timeout to be bigger, knock the maximums
down in these constraints. I don't believe that a longer-running test
will show any new types of behaviours, so we can always increase the
reseed count (and run the tests in parallel!) if we're not covering
enough of the state space.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>