[dv] Fix a race condition in tl_device_seq
This condition came about because the pre_start task for tl_device_seq
clears a flag, which doesn't work if we set the flag just before
starting the sequence.

That clearing behaviour came from commit 17d386c879a, which doesn't
really explain why it's needed. Maybe we can drop that behaviour
entirely, but there's an easy work around just in case it's necessary.
This commit does that workaround.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>