[sram_ctrl/dv] Avoid updating readback CSR on key_req
When a key request is triggered by the DV environment, we
shouldn't update the readback CSR as this triggers the
error described in lowRISC/opentitan#23445.

More specifically, writing the readback CSR over the RAL
interface issues a TL-UL message to the register interface
of the sram_ctrl module. However, when cf.in_key_req=1, the
sram_ctrl_scoreboard checks that no TL-UL transaction is on
the bus.

By waiting until cfg.in_key_req we avoid this error message.

Closes #23445.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>