[doc] Add V2S scoping to V2 checklist
The existing V2 checklist had an item (V3_CHECKLIST_SCOPED) to
understand scope and estimate effort for V3. We added the V2S
development stage since that was defined, and we're definitely going
to need to estimate effort for that at some point!

Add the task to the V2 checklist accordingly. It will be a reasonably
quick job for most blocks.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>