[doc] Add V2S scoping to V2 checklist
The existing V2 checklist had an item (V3_CHECKLIST_SCOPED) to
understand scope and estimate effort for V3. We added the V2S
development stage since that was defined, and we're definitely going
to need to estimate effort for that.

This commit splits things so that the V2 signoff process includes
scoping for V2S and also for the functional verification parts of V3.
The "full scoping" for V3 moves to the V2S signoff process.

This change to the checklist was discussed in the TC meeting on 4th
June 2026.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>