[rom] Change the endianness of the digest for the HMAC driver.
Previously, the HMAC driver emitted the digest in "little-endian" mode;
this change flips it to "big-endian". Normally, we represent everything
in ROM in little-endian, but in the case of HMAC/SHA2 the output is more
like a byte sequence than an integer.

Standards love big-endian, so both ECDSA and RSA interpret the byte
sequence as a big-endian integer. So, when we only needed those,
little-endian kind of made sense; we produce the output with flipped
bytes and then just flip them again when we interpret the integer in the
way OTBN and most other processors prefer (little-endian).

However, the second you want to use a digest as anything other than an
integer (for example, in SPHINCS+), the flipped bytes cause issues. And
given the choice, it makes more sense to me to represent the bytes in
the same order on the processor as the SHA2 specification dictates,
rather than in the order such that they interpret to the right integer
for asymmetric crypto specifically.

Signed-off-by: Jade Philipoom <jadep@zerorisc.com>