[otbn,dv] Fix wait_for_run_completion
I actually broke this logic with commit f70443076e0. Before that
commit, we waited until the status was something different from the
"run command" that we expected to be running. In this commit, I
changed the logic so that it snooped the current status and waited
until the status had changed.

Of course, I might do the snooping too late and then we end up waiting
until the status is no longer idle. That times out... This commit
changes things so that we wait until we're Idle or Locked (so no
longer running one of the three commands).

The timeout (that already existed in the polling case) is now
explicitly encoded here, and applies to the interrupt case as well.
I've made it 1ms pretty much arbitrarily. We can always bump it up if
it is causing things to fail.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>