[otbn,dv] Avoid extra wait after the end of a run in otbn_base_vseq
This logic was added with an old version of the code where we were
waiting until status changed, then had to wait until the post-run
secure wipe had finished.

In the last commit, I changed the initial wait so that it waits until
we're Idle or Locked. As a result, this second wait will always finish
instantly. Remove the unnecessary wait.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>