[test] Add pwrmgr low power test with pending IRQ
Add pwrmgr test that checks for correct recovery of the CONTROL CSR when
Ibex does not issue a long enough pulse for pwrmgr to sample a core
sleeping state. This occurs when an IRQ is pending before entering WFI,
and the prefetcher is active when WFI is executed.

Software should be able to clear LOW_POWER_HINT when pwrmgr does not
begin the low power entry sequence.

(cherry picked from commit c7e72a6)

Signed-off-by: Alexander Williams <awill@opentitan.org>
Co-authored-by: Edward Hill <ecgh@google.com>