[otbn,dv] Increase run timeout by a factor of 10 to 10ms
The previous timeout was 1ms, which I chose pretty arbitrarily. It
turns out that this isn't long enough for some runs in the otbn_multi
vseq.

Increase to 10ms, which is still short enough that a failure should
only take a couple of minutes.

This commit also tweaks the timeout error message. When I originally
wrote it, it only applied to status polling timeouts. Before pushing
the PR, I generalised things to cope with interrupts too, but forgot
to fix the error message. Now done!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>