[ci] Introduce a test orchestrator written in bazel
The commit fundamentally changes how tests are considered for running
in CI. Previously, any test not marked as "manual", "broken" and "skip_in_ci"
would be run in CI in all execution environment for which it is defined. We
then came up with some clever bazel queries to try and avoid running them
in multiple execution environment when not necessary but this is quite fragile.

The approach is very different: when creating the tests using opentitan_test,
we call a new function, called "ci_orchestrator" with the list of all exec_env
for which it is defined. This function then returns a subset of that list that
should be marked with skip_in_ci. The current implementation is that there is
a list of exec_env that are mutually-exclusive (_ONLY_RUN_ONE_IN_CI_SORTED)
and all but one will be skipped: the only one to run will the highest one in
the list. With this approach, we are guaranteed to never run duplicate tests.
If more complicated logic is required, it can always be added to this function.

As a byproduct of this, we can significantly simplify the filtering logic when
running FPGS jobs.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 3dedc47374d85691e7d16be897e47da5bf975407)