[driver, hmac] Use `Idle` bit to poll HMAC
Previously HMAC was waiting `hmac_done`
interrupt to deduce that ongoing operation
was completed.

This commit changes this behavior by using
`hmac_idle` bit of STATUS instead. The idea
is to increase the test coverage for this flag.
`hmac_done` interrupt is still read afterwards
to check that both correctly indicate the end
of an operation

Signed-off-by: Fatih Balli <fatihballi@google.com>