[bazel] Make device test exec envs relative to EARLGREY_TEST_ENVS
Also `EARLGREY_SILICON_OWNER_ROM_EXT_ENVS`.

Basically, we want as many tests as possible to use `EARLGREY_TEST_ENVS`
so there's a single place we can control the environments we test on.

Where a test must not have an environment because it's not designed to
run there, we use `dicts.omit` to remove it.

Where a test should run in an environment but currently doesn't because
it's broken, we tag that environment as `broken` so it's not run in CI.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>