[prim/rtl] Allow for tech-specific impls of `prim_flop_2sync`
This allows for technology/library-specific implementations of
`prim_flop_2sync`.  The generic implementation is based on two
back-to-back `prim_flop`s, which is identical to how `prim_flop_2sync`
was implemented before.

This commit doesn't add FPGA-specific implementations; this will be
tracked in an issue and might get implemented later.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>