[exec_env] Fix commandline arguments
The `--rcfile=` argument in the `silicon_owner_sival_rom_ext` is
incorrect.  That argument is used to control or override the loading of
`$XDG_CONFIG_HOME/opentitantool/config`, which is a list of command-line
arguments the user wants applied to each invocation of `opentitantool`.

It cannot parse a json config file (that can be done with `--conf=...`).
In this case, using `--conf` should not be necessary because any of the
built-in configurations are resolved and parsed by the `--interface`
selection.

Signed-off-by: Chris Frantz <cfrantz@google.com>