[otbn,dv] Weaken check in otbn_trace_entry.cc
The existing code wanted to see exactly two writes to each register in
the RTL, and then check that the second write had the same value as
the model expected.

This is slightly too tight, because we want to allow situations where
some other rubbish gets written to the register in the meantime.
Weaken the check so that we require at least two writes, and have
predictions about the final write (as before).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>