[csrng] Make reseed counters readable at any point in time
Previously, the reseed counters were only readable if the
otp_en_csrng_sw_app_read fuses were enabled and if the
CTRL.READ_INT_STATE field was enabled, i.e., if the full internal state
was readable.

However, the reseed counter should be readable even if the internal
state is not readable because this allows firmware to track the age
of the different CSRNG contexts. The reseed counter itself is not
sensitive information as it doesn't allow inferring the values of
Key and V.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>