[i2c,dv] Add a more-precise host-mode performance test
This commit adds a new vseq and a monitor task that enables the precise
monitoring of the bus performance operating in DUT-Controller mode.

The monitor task perf_monitor() captures the SCL period for each data bit in a
transfer, and pushed them to a queue in the agent cfg. The vseq can then
'subscribe' to this queue, and check each bit period against it's expectation.
The monitor needs to judiciously enable the monitoring code when clock
stretching is not expected, nor clock synchronization. In this use-case, we can
form a precise expectation for the period.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>