[flash_ctrl/rtl] Host data XOR address infection
This commit adds the data XOR address infection mechanism. The goal
of this countermeasure is to detect faults on read addresses targeting
the flash controller.

On a host flash read, the address used for accessing the data is XORed
inside the flash_phy_rd module to the fetched data.

The address then is removed in the u_tl_adapter_eflash where the core has
a TL-UL interface to the flash memory. Here, the buffered address from the
incoming TL-UL read request is taken.

When a FI attack manipulates an address, data would be fetched from the
wrong address in the flash memory. But inside the flash_phy_rd module,
this manipulated address is XORed to the data. Now, when removing the
address by using the original address, and there is a difference from
a FI, the data would be faulty and the subsequent integrity check detects
the attack.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>