[sw/host] Fix UART usage regarding control flow
These rust harnesses enable control flow on a UART object but then
immediately drop it and gets it again in each test function. Since
the UART object is reference counted, it will be destroyed in-between
tests and the control flow setting will be lost. The easiest fix is to
create the object once and pass it around.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>