[otbn,dv] Weaken timeout in otbn_rf_base_intg_err_vseq
The existing 20,000 cycles count was pretty arbitrary and I picked it
based on a time that seemed to work. But it's a bit rubbish because it
takes the EDN timing into account (we might be waiting on a random
seed).

Change things so that we're now waiting until the current execution
finishes. We might still be unlucky though: start_running_otbn might
have ended on the last couple of instructions in the run. On a quick
check, that happened 3 times in 100 runs. To make that much more
unlikely to kill the test, we repeat up to 4 times, so I'm hoping that
the probability of a timeout for that reason is now less than one in a
million (since 0.03^4 = 8.1-7).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>