[otbn,dv] Disable prim_fifo_sync assertions in some FI tests
The assertoff/asserton pair gets a bit more general than the existing
code, to work around the fact that you can't pass an arbitrary
expression to $assertoff/$asserton, so there's not a nice way to
disable just the correct one without lots of code repetition.

This change is needed because the imem's u_sramreqfifo is newly a
prim_fifo_sync and we need to disable its integrity assertions if we
corrupt it.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>