[prim_fifo_sync/rtl] Fix err_o behavior
When an err_o is raised by the u_fifo_cnt (there is a ptr error
inside this module), we should not raise the rvalid_o signal of
the FIFO and we should not forward rdata_o = rdata_int as this
data is not valid.

Instead, set rdata_o = '0 and deasser the rvalid_o signal.

Closes lowRISC/opentitan#23578

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>