[otbn,dv] Generalise RMA request tracking in model
There's not a big difference here, except that we now keep track of
"the current state of rma_req_i" instead of reacting to an event that
says "rma_req_i become ON".

This should allow us to teach the Python model to only react to the
value of rma_req_i at appropriate times.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>