[flash_ctrl/dv] Fix sec_cm assertions
This commit disables assertions that get triggered during
sec_cm assertions. Disabling these assertions is valid
as, on a FI, the u_sramreqfifo, u_rspfifo outputs could
be manipulated. Although this error is captured by
intg_error_o the current flash DV env is not aware of this
and triggers invalid data assertions.

Moreover, we need to turn off the flash_access_after_disabled
function when inducing faults into the counters of the read
FIFOS. The reason for that is:
- A fault in these FIFOs could manipulate a TL-UL message
- This is detected by the DV environment by sensing the fatal
  alert
- If we would now call the flash_access_after_disabled function
  while there still is a manipulated TL-UL message in the system
  we could get a timeout
- Instead, we shouldn't rely that we still can access the flash
  after disable when there is a fault injection attack going on

Partly addresses #23567, as the `sram_ctrl` DV environment needs
to be adapted similarly.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>