[rv_dm] Fix RDC and DFT issue with resets
Convert the JTAG combined resets to have a synchronous assertion timing
in the core clock domain, placing them on the D input of the first flop
of the prim_flop_2sync, instead of the reset inputs of both flops. This
removes the asynchronous reset assertion that may cause RDC issues, and
it restores scan reset as the only reset for those 2 flops.

Note that prim_fifo_async_simple guarantees that a prior transaction in
flight will at least maintain its data values, since the holding
register is *not* reset. This mechanism does not guarantee that the
request will be squashed, but it *does* guarantee that the request will
not be malformed by the reset. In addition, if the JTAG reset is held
for long enough (i.e. >= the round-trip delay), the response for any
previously in-flight request will be dropped.

Signed-off-by: Alexander Williams <awill@opentitan.org>