[otbn,dv] Disable FI checks in DMEM req fifos
These currently cause spurious errors in the otbn_sec_cm test. The
problem is that the count in the request fifo gets forced from 0 to 1.
This spits out a bogus TL request, and also causes an alert to be
raised by the sram adapter.

The alert is good: it's exactly what is supposed to happen.
Unfortunately, the request that goes out might be X (because the fifo
has never not been empty). Unfortunately, the TL response ends up
polluted with X and this ends up polluting the start/stop control FSM,
causing it to get stuck in the X state.

The end result is that we see a fatal alert with no corresponding
status change, causing a DV check to fail (because we always expect
OTBN to lock up after it generates a fatal alert).

This change disables fault injection on these particular FIFO
counters, side-stepping the issue.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>