[sram_ctrl,dv] Disable FI checks in tlul_adapter req fifos
These currently cause spurious errors in the sram_ctrl_sec_cm test. The
problem is that the count in the request fifo gets forced from 0 to 1.
This spits out a bogus TL request, and also causes an alert to be
raised by the sram adapter.

The alert is good: it's exactly what is supposed to happen.
Unfortunately, the request that goes out might be X (because the fifo
has never not been empty). Unfortunately, the TL response ends up
polluted with X and this lets multiple TL-UL integrity checks within
the sram_ctrl DV environment fail.

This change disables fault injection on these particular FIFO
counters, side-stepping the issue.

Based on lowRISC/opentitan#23627

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>