[opentitantool] Allow manifest updates with ELF info
Adds the --elf flag to the `image manifest update` subcommand that takes
the corresponding ELF file for the binary being modified as input. When
supplied, opentitantool will populate the `code_start`, `code_end` and
`entry_point` fields in the manifest.

This is useful when building tock firmware, since defining the manifest
as a static struct as in the ROM_EXT doesn't work as nicely in rust.

Signed-off-by: Jon Flatley <jflat@opentitan.org>