[pattgen,sival] Update pattgen_ios to support inactive levels
Since inactive levels require a few changes to the test logic, this
commit makes a few more refactoring. The biggest change is that on
failure, the code will output a VCD dump of the recorded waves
and the expected waves, which makes it much easier to understand
the discrepancies.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>
(cherry picked from commit 790f3ce3e02c24004121a95585706010a0896763)