[tl,dv] Remove check from tl_host_driver.sv
This is in a task that collects TL responses and uses them to respond
to the TL item that made the request (calling put_response on
seq_item_port).

It gets a bit surprised if we get a response with no request! Disable
the check (there's a detailed explanation of why this is the right
behaviour in a comment there).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>