[csrng/rtl] Make the main SM error test signal fatal
The main SM error test signal is currently the only test signal where
the actual error would be fatal but the test signal is not.
This commit make adds the err_code_test_bit for the main SM error to
fatal_loc_events to make the main SM error test fatal.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>