[csrng/dv] Add the right err_test codes to fatal errors
For each error there is a corresponding error test signal.
In the error test sequence not all of these error test signals
were expected to cause a fatal error even though they do in
the normal case. This commit adds for each fatal error, the
corresponding err_test to the list of fatal errors.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>