[otbn,dv] Don't call _abort in state.py on RMA request
The _abort function is for rolling back the changes that a cancelled
instruction was trying to make. This isn't really needed for RMA
request modelling. In the normal use case for RMA requests, there
isn't an instruction running because we haven't actually started an
operation yet.

Indeed, RMA requests during an operation are actually in the RTL
(although we don't model that at the moment - a job for a later
change).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>