[dv] Reduce number of seeds for stress_all_with_rand_reset tests
This is a V3 test and quite a few IP blocks for which we are primarily
targeting V2S (AES, CSRNG, HMAC, I2C, KMAC, RV_DM, USBDEV) have a very
low pass rate for this test right now.

To save DV resources, this commit reduces the number of seeds for this
test for blocks not targeting V3 right now.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>