[tlul] Avoid inference of latch
The `addr_xor` signal defined inside an always_comb block did not have a
default assignment. In particular, if `sramreqfifo_rdata.mask` was
all-zero, it was undriven. This should be avoided as tools may infer
latches. This commit slighly simplifies the code not require the
`addr_xor` signal in the first place.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>