[csrng/dv] Fix sfifo_gadstage_err and sfifo_gbencack_err tests
The csrng_err_vseq was failing for sfifo_gadstage_err.
This was due to some undefined bits that were propagating through
CSRNG after a signal is forced. The forced signal caused a FIFO to
push in some undefined bits. This caused assertions to fail and
eventually the tlul interface to throw an error due to unexpected
data. This commit fixes the above issue by forcing all of the input
bits of the downstream FIFO to zero. This way no undefined bits will
propagate.

Something similar happens for the sfifo_gbencack_err. This commit
also fixes the sfifo_gbencack_err test by forcing adstage_glast to 0.
This ensures that no request to the downstream arbiter is made, which
used to trigger some assertions.

Signed-off-by: Hakim Filali <hfilali@lowrisc.org>