[otbn,dv] Correct the logic about next STATUS at end of wipe round
The existing DV code jumped to the end of a secure wipe after a single
round if we saw an RMA request in the middle of that first round. This
isn't what the RTL does, and doesn't really match what we want.

On the last wipe round, we were also using the current "wiping type"
to figure out whether to jump to IDLE or LOCKED. This is not correct
if the RMA request is asserted in the middle of the wipe. Honestly, I
probably shouldn't have separated WIPING_GOOD and WIPING_BAD. Maybe
they would be best to merge again, but this commit is simpler and just
checks for the RMA signal at the right time.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>