[lint] Waive USE_BEFORE_DECL AscentLint errors for SV functions
We typically define SV functions towards the end of package or module
files in our code base. Latest versions of AscentLint flag a
USE_BEFORE_DECL error in this case.

For signals, this could indeed be an issue (at least some simulation
tools don't cope with that) but to the best of my knowledge I don't
think there are issues for functions. It rather seems common practice
to define them towards the end of files. This commit thus waives these
errors for functions where the tool flags an error.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>